corrected policies

+++ app/Http/Controllers/FollowTagController.php
@@ -12,17 +12,18 @@ use PDOException;
 class FollowTagController extends Controller
 {
     /**
-     * Follow or unfollow a tag
+     * Follow a tag
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return JSON The followed tag content in JSON format.
      */
     public function follow(Request $request)
     {
         $user_id = Auth::user()->id;
         $tag_id = $request->id;
 
-        //$this->authorize('delete', $tag); //TODO: follow policy
+        $this->authorize('follow', FollowTag::class);
+
         $follow_tag = new FollowTag();
         $follow_tag->user_id = $user_id;
         $follow_tag->tag_id = $tag_id;
@@ -40,16 +41,17 @@ class FollowTagController extends Controller
      * Unfollow a tag
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return JSON The unfollowed tag content in JSON format.
      */
     public function unfollow(Request $request)
     {
         $user_id = Auth::user()->id;
         $tag_id = $request->id;
 
-        $follow_tag = FollowTag::find(['tag_id' => $tag_id, 'user_id' => $user_id]);
+        $follow_tag = FollowTag::findOrFail(['tag_id' => $tag_id, 'user_id' => $user_id]);
 
-        //$this->authorize('delete', $follow_tag); //TODO: unfollow policy
+        $this->authorize('unfollow', FollowTag::class);
+
         try {
             $follow_tag->delete();
         } catch (PDOException $e) {
